[Jun2018] ca2 store application
bilbo and Mummi (Léia Maheli de Souza Tsumanuma contribution, http://bisba.com/! A melhor mãe!!) contribution!! veriverse contribution!! veriverse backyards people lol and talk contribution!! veriverse neighborhood contribution!! https://LiveEdu.tv/ (Livecoding.tv) contribution!! ThomasBS (Thomas Boregaard Sørensen, https://kaem.dk/, https://kaem.io/), drmjg, camwhite (Cameron White, Twitchs' scamwhite), xmetrix, intelijens, syaseigirl, CatherinaHelena (CatherinaCortega, Twitchs' helenawinter), fuel, r00tkiller, hiiamgosu (itashellhaha...(gosu), scorpioinyouranus... ), pablosj (sjpablo, Pawel Kulesza, Pablosj special contribution (after so much days with so low audience... ) .. contribution!!), phanxgames, dunno_aye (Dunno_Aye, Andreas from Denmark from Twitch contribution!!), 0din, 7ru57, anita, aqumb, beater, benjones, bill_gates, brayoeh, camilosasuke (me :ca2:), cary, chen68, clementin, Cookiesliyr, costinis, crisy, danixo, dauntlesstech, david_joffe, davids81, denvercowboy, dmitryb, drcode, dslegends, ederangelo21..., epryka, evgeniy_andreev, famed, flythiago, frankly444, georgeb, gigaherz, gostrafx, grahambell, guacjack, hatyme, hbw, heinz3, hellomike, hishritaha, hosjenken, intelijens, irwanbd, isabella, jackperrie, jacobs, jaycee, johnnydully, kaeris, kamisama, kannu, karrmarr, kayne3, Kennis, kissloryshy, klerind, kozmo, kranalex, kruzoe, kurtig, landry96, lemonay, lex_yang, larrybrook, lilbro1062000, LiveEdu_bot, liveedu_support, LiveOrDevTrying, LuisCauro, lyndroid, maoabc1818, mark_miller, masterchoc, matroskinb, mitch_ill, mitsurugi, nero951, ngetich, nightfaust, nilnala, nini520, noeldavies, novicehere, olivert, olelukoe, ong, pestka12, pinkgorilla, pinky, pirateofthesea, polodonald, rex2, resident, riverside, rooney2, sale (Serbian?), samdev, sargin, shgalym (from Kazakhstan), shieldgenerator7 (Twitchtv shieldgenerator7), shinkaisokuuindo, siox, starlord91, stefan_v, stephen22, systemis, teleporter, thermonucleary, thiagovgamg, tonycode, valdinei, vapaamies, vodkavodkavodka228, xianmin12, wariosolis, wetriks, wilson2000, yanaseattle, zakaria98, zzeroge and zzzel contribution!! https://Twitch.tv/ contribution!! ! ThomasBS_ (LiveEdu.tv ThomasBS), RocketBeans.TV (Nils, Simon, Colin, Donni, Etienne, Florentin, Lars, Marco, Oliver, Renè, [Schröck, Daniel], Renè, Timo, Budi und Viet), AlfredoPlays (and his mates... Alfredo, Hanson, Rey and Mackie contribution), KindaFunnyGames (Greg Miller, Nick Scarpino, Andy and Kevin contribution...), handmade_hero (Casei Muratori contribution...),  SoulFoamTV (LettuceSmokeStudios, SoulFoam), scamwhite (Cameron White, Liveedutvs' camwhite, mesh.io, rivanov23 contribution, unicorn_dev (Liveedu's unicorn? rosen?)...!), Phanxgames (Live's edu phanxgames, Henri Price contribution...), Peter_Alfons_Loch (from phanxgames stream and discord channel, first "explicit" and "spontaneous" commit and contribution to this newly create GitHub repository), cancerMoonLife(hiiamgosu from LiveEdu, xxxKLEOPATRAxxx, brumbrum9000 contribution), Spoofh (LiveEdu.tv Fuel: Benjamin Kuhl), themushmouth (LiveEdu.tv mushmouth: Patrick McFarling), Tectuma (Kevin, Stacie (GothicCat1), Matt Brock (Cronos) and Dylan and Nick...), IILomoo (Liveedu lomoo, Matti), pewteeweet (liveedus's hailviral1),nothings2 (Sean Barret), zakaMitchJones (Nick [Nicholas], Greek, sodappoping contribution), iDunno_XD (Andreas, ColdessRotMG, ColdnessRotMG, Dunno_Aye), CantorSys (VapaamiesTV, Vladislav Javadov, liveedu's () contribution...), sundz98, sonnystar (hairy green sonny star, Vlad Ionescu), WildN00b (Wild from LiveEdu/LiveCoding), RobTheSwan (and Teazal cat contribution... ), chrif_rachid (Chrif Bouderka Rachid), ReaLomo (Lomoooooo, Matti), tarikogut (Liveedu's tarikogut), Wheeze202, helenawinter (CatherinaHelena, CatherinaCortega from LiveEdu), userpan (livecoding, liveedu's epryka, Vova, Vladimir contribution...), harmannieves (Harman Grewal), XanderGamingTwich and xmxstudio (LiveCoding/LiveEdu xmetrix, and his Friends, Beuca, unicorn_dev), 03F001 (and C++ Twitch Community contribution..), eduubessa (from Liveedu eduubessa), Naysayer88 (Johnanthan Blow), Shmellyorc, ericswenson15(16) (Eric Jonathan Swenson), ws_ubi, pedropcruz, Serpent_AI (contribution and their raided contribution: Kitboga, MikeSci and Skylias contribution...) , 2hmedia (Matthew and his partner contribution...), zakariawaseem (Liveedu's zakaria98 contribution...), HonestDanGames, Adan_204, Adevel, agathonaooo, agent_zero009, albertisufi, Angel37558, angelawuaw, Always_Betonentei, Beuca, BlavenX, Bromadir, Bugrimov_Alex, CakeBiscuit_, CogWhistle, CommanderRoot, COOKIEVIRUS, CuteHealGirl, Cylverrose, david_joffe, DeadLikeImEarnHardT, Derteddii, dimer525, DrewHookah, dzyrr_tv, electricallongboard, electricalskateboard, emielrregis, Evgeniy_andreev, faegwent, faillehad (french swearer with igrectepe) ), Fataldizz, foldm, Fonz_tv, Funmartas_cz, GameDuells, Gaseline, GiveMeHoneyNow, gnomer0, Grononmc, guy_ross, hazurl, MasterTay, igrectepe (French swearer with faillehad), itechno2016 (Russian guy), J0ko, Javo_net, Jd_0950, JuniorRoy, Junmakii, katyatavalika (from Livecoding), KingNoseBleed, Kingtykari, kissloryshy, kkkae, knutaf, Krassenstein, LANFusion, Lauheid1974, Lessio__, LiveOrDevTrying, lingpling, local_root, LoryMaster, Luxio_tv, MatejKunc, megadamage, Miouchfr, Mopynatv, Mrxp3r7, multiparadigma, nexuslitch, nightbot, OrcaPrime, OxenHill, Paolo435, Piepassaro, Pixel701, Pucki4, Puka420, recursionishard, Restreamiobot. Retched, RidixCr, RoboRock81, Rqndy1, Sassafras_wot, ScorpionIlluminati, shieldgenerator7 (Liveedutv shieldgenerator7), ShoutGamers, SirKenta, SirKultan, Skairyhairy, Skorp_NY, SlayerXen, Slocool, Smtryu, solve_accept, Stannys88, Stay_Hydrated_Bot, Stevoid1990, stopplayer, sumberg, SuperCoolNinja1, supergaling (Thomas Johansen), SystemOfPeter, Tanerax, The_eQual_, theformemrcc, TheInPrisoned, thmca, thronezilla, Tookietookie, topbraj, Totizera, TrollLolVin, Trump88, TutorExilius, Vistiisen, Volon_, Vulsen, Worfox, Wraithone and XekeDeath contribution!! https://mixer.com/ (previous Beam.pro) contribution!! R4ver contribution (and his friends agile as him, from Sverige... contribution), ArrestedMoney2, AtomicNinja, Blez, GreenGully, Kestalkayden, leiamaheli, LlamaJuic3467, ModestTim, NormalZeus53319, OurTwinLife, sharkshooter641, SpartanUmaw, TaxingBinkie154, TopazRickC and WedRhombus23 contribution!! YouTube Live contribution!! resident ca2 (resident@ca2.email), Desenvolvimento 3 Mediaplus (thiagovgamg from LiveEdu), Kalash (the IT Passionate), NzLeoGenX X, Shivam Bhola contribution!! bisba contribution!! Carol, Lúcio and Lucinho contribution!! Upwork.com contribution!! Bock Engineering contribution!! Infinisys and Dennis Heazle contribution!!

from bilbo at veriverse!! Internet Net Curitiba 120Mbps▼ 10Mbps▴
i7 6700 Dell XPS 8900 32GB DDR4 RAM 2x240GB 1x480GB SSD 2TB HDD
Windows 10 Pro (Octo Boot with macOS High Sierra 10.13.5, openSUSE Tumbleweed, Windows 8.1 Pro, Ubuntu 18.04, Fedora 27, Manjaro Linux 17.1.10 and Windows 10 Pro with Hyper-V)

+++ appseed/aura/filesystem/filesystem/filesystem_dir_system.cpp
@@ -780,7 +780,7 @@ namespace file
       bool system::is_dir_map::lookup(const ::file::path & path, bool &bIsDir, u32 & dwLastError)
       {
 
-         return lookup(path, bIsDir, dwLastError, path.length());
+         return lookup(path, bIsDir, dwLastError, (i32) path.length());
 
       }
